treewide: remove unnecessary git-compat-util.h includes in headers
Since git-compat-util.h needs to be included first in C files, having it
appear in header files is unnecessary.  More importantly, having it
included in header files seems to lead to folks leaving it out of C
files, which makes it harder to verify that the rule is being followed.
Remove it from header files, other than the ones that have been approved
as alternate first includes.

Signed-off-by: Elijah Newren <newren@gmail.com>