config.c: create config_reader and the_reader
Create "struct config_reader" to hold the state of the config source
currently being read. Then, create a static instance of it,
"the_reader", and use "the_reader.source" to replace references to "cf"
in public functions.

This doesn't create much immediate benefit (since we're mostly replacing
static variables with a bigger static variable), but it prepares us for
a future where this state doesn't have to be global; the "struct
config_reader" could be provided by the caller, or constructed
internally by a function like "do_config_from()".

A more typical approach would be to put this struct on "the_repository",
but that's a worse fit for this use case since config reading is not
scoped to a repository. E.g. we can read config before the repository is
known ("read_very_early_config()"), blatantly ignore the repo
("read_protected_config()"), or read only from a file
("git_config_from_file()"). This is especially evident in t5318 and
t9210, where test-tool and scalar parse config but don't fully
initialize "the_repository".

We could have also replaced the references to "cf" in callback functions
(which are the only ones left), but we'll eventually plumb "the_reader"
through the callback "*data" arg, which will allow us to rename "cf" to
"cs" without changing line lengths. Until we remove "cf" altogether,
add logic to "config_reader_*_source()" to keep "cf" and
"the_reader.source" in sync.

Signed-off-by: Glen Choo <chooglen@google.com>