fsmonitor: handle differences between Windows named pipe functions
CreateNamedPipeW is perfectly happy accepting pipe names with seemingly
embedded escape charcters (e.g. \b), WaitNamedPipeW is not and incorrectly
returns ERROR_FILE_NOT_FOUND when clearly a named pipe, succesfully created
with CreateNamedPipeW, exists.

For example, this network path is problemmatic:
\\batfs-sb29-cifs\vmgr\sbs29\my_git_repo

In order to work around this issue, rather than using the path to the
worktree directly as the name of the pipe, instead use the hash of the
worktree path.

Signed-off-by: Eric DeCosta <edecosta@mathworks.com>