hash.h, repository.h: reverse the order of these dependencies
Previously, hash.h depended upon and included repository.h, due to
the definition and use of the_hash_algo (defined as
the_repository->hash_algo).  Move this #define, and the associated
inline functions that depend upon it, from hash.h to repository.h.
Due to that movement, reverse the dependencies so repository.h includes
hash.h.

This allows hash.h and object.h to be fairly small, minimal headers.  It
also exposes a lot of hidden dependencies on both path.h (which was
brought in by repository.h) and repository.h (which was previously
implicitly brought in by object.h).

Signed-off-by: Elijah Newren <newren@gmail.com>