config.c: introduce kvi_fn(), use it for configsets
In order to pass "struct key_value_info" to config callbacks in a future
commit, we want to make sure that a "struct key_value_info" is available
in all code paths that invoke a config callback.

Enforce this by introducing a helper function, "kvi_fn()", that sets
"the_reader.config_kvi" before invoking the config callback, and
trivially use it in "configset_iter()". This closely simulates passing
an extra "struct key_value_info" such that making the switch will be
trivial.

This breaks our rule that no underlying machinery uses "the_reader", but
"struct config_reader" will be gone by the end of the series so there's
no point plumbing it as an arg.

Signed-off-by: Glen Choo <chooglen@google.com>