git-commit: fix cut_line behavior with CRLF
On Windows with certain editor (mis)configurations, the cut_line that is
used when commit.verbose is turned on doesn't work correctly because
COMMIT_EDITMSG is rewritten with Windows line endings. This used to
happen on old versions of VSCode, but I can't seem to reproduce it
anymore without explicitly setting the line-endings to Windows. Even so,
I think it would be sensible to fix this for good, even in the face of
bad editor behavior.

This patch moves the \n ending out of the detection string constant, so
that regardless of the line endings, the cut_line is detected correctly.

Signed-off-by: Robert Boyd III <git@robertboyd.dev>