(RFC-only) config: apply cocci to config_fn_t implementations
Pass "struct key_value_info" to *most* functions that are invoked as
"config_fn_t" callbacks by applying
contrib/coccinelle/config_fn_kvi.pending.cocci. None of the functions
actually use the "kvi" arg yet (besides propagating it to a function
that now expects "kvi"), but this will be addressed in a later commit.
When deciding whether or not to propagate "kvi" to an inner function,
only propagate the "kvi" arg if the inner function is actually invoked
elsewhere as a config callback; it does not matter whether the function
happens have the same signature as config_fn_t.

This commit does not build and has several style issues (e.g. a lack of
spacing around the "kvi" arg), but I've split this out for the RFC so
that it's more obvious which changes are automatic vs manual. Post-RFC I
will squash this with the rest of the refactor + cocci changes.

Signed-off-by: Glen Choo <chooglen@google.com>