[RFC] transport: add --show-service option
Without using protocol v2, the git server needs to send a pktline
"# service=$servicename" to the git client first. This often
requires the git server to implement it independently, but it can
be delegated to the `git receive-pack` and `git upload-pack` to complete
the work proactively. Therefore, the `--show-service` option is added
to `git receive-pack` and `git upload-pack`, which can be used to send
the "# service=$servicename" pktline, making the logic of the git
server more concise.

Note that this `--show-service` option can only be used together with
`--http-backend-info-refs` and it is not applicable when using protocol v2.

Signed-off-by: ZheNing Hu <adlternative@gmail.com>