[RFC] transport: add --show-service option
Without using protocol v2, the git server needs to send a pktline
"#service=$servicename" to the git client first. The servername
here can only be selected in "git-receive-pack and "git-upload-pack",
neither the git client nor the server depends on this servername to
perform any functions. Howerver, , implementing this capability
independently can be cumbersome for the git server, as seen in [1].

To simplify this process, the `--show-service` option was added to
git-receive-pack and git-upload-pack. This option can be used to
send the '#service=$servicename' pktline, making the logic of the
git server more concise. Note that this option can only be used
together with --http-backend-info-refs and is not applicable when
using protocol v2.

[1]: https://gitlab.com/gitlab-org/gitaly/-/blob/master/internal/gitaly/service/smarthttp/inforefs.go#L82

Signed-off-by: ZheNing Hu <adlternative@gmail.com>